+++ components/adapters/simulator/src/simAdapter.ts
@@ -29,15 +29,29 @@ export class SimulatorAdapter extends core.Adapter {
   }
 
   public async deploy(): Promise<core.DeployResult> {
+    const envs: Record<string, string> = {
+      PLUTO_PROJECT_NAME: this.project,
+      PLUTO_STACK_NAME: this.stack.name,
+      PLUTO_LANGUAGE_TYPE: this.language,
+      PLUTO_PLATFORM_TYPE: this.stack.platformType,
+      PLUTO_PROVISION_TYPE: this.stack.provisionType,
+      WORK_DIR: this.stateDir,
+    };
+
     this.simulator = new Simulator(this.rootpath);
-
     await this.simulator.start();
-    process.env.PLUTO_SIMULATOR_URL = this.simulator.serverUrl;
+    envs.PLUTO_SIMULATOR_URL = this.simulator.serverUrl;
 
-    await this.simulator.loadApp(this.archRef);
+    for (const [key, value] of Object.entries(envs)) {
+      process.env[key] = value;
+    }
+
+    const outputs = await this.simulator.loadApp(this.archRef);
+    const awaitedOutputs = await awaitNestedPromises(outputs);
 
     return {
       outputs: {
+        ...awaitedOutputs,
         simulatorServerUrl: this.simulator.serverUrl,
       },
     };
@@ -59,4 +73,15 @@ export class SimulatorAdapter extends core.Adapter {
     data;
     throw new errors.NotImplementedError("The simulator adapter cannot be reused.");
   }
+}
+
+async function awaitNestedPromises<T>(obj: T): Promise<T> {
+  for (const key in obj) {
+    if (obj[key] instanceof Promise) {
+      obj[key] = await obj[key];
+    } else if (typeof obj[key] === "object") {
+      obj[key] = await awaitNestedPromises(obj[key]);
+    }
+  }
+  return obj;
 }