+++ packages/pluto-py/pluto_client/queue.py
@@ -8,6 +8,7 @@ from pluto_base.resource import (
 )
 from pluto_base.platform import PlatformType
 from pluto_base import utils
+from .utils import create_simulator_client
 
 
 @dataclass
@@ -65,5 +66,10 @@ class Queue(IResource, IQueueClient, IQueueInfra):
             from .clients import aws
 
             return aws.SNSQueue(name, opts)
+
+        elif platform_type == PlatformType.Simulator:
+            resource_id = utils.gen_resource_id(Queue.fqn, name)
+            return create_simulator_client(resource_id)
+
         else:
             raise ValueError(f"not support this runtime '{platform_type}'")