+++ packages/pluto-py/CHANGELOG.md
@@ -1,5 +1,11 @@
 # pluto-client
 
+## 0.0.14
+
+enable Python execution in simulator
+
+This commit introduces the ability to execute Python within the simulator. The resource infrastructure SDK implementation has been updated to adhere to the IResourceInfra interface, unifying the implementation for both local and cloud environments. Additionally, the Website, Tester resource type is now supported in the simulator environment.
+
 ## 0.0.13
 
 feat(sdk): add Vercel deployment for ReactApp and Website